+++ apps/web/src/app/billing/organizations/organization-subscription-cloud.component.html
@@ -79,17 +79,6 @@
                 {{ i.quantity * i.amount | currency : "$" }} /{{ i.interval | i18n }}
               </td>
             </tr>
-            <tr bitRow *ngIf="discount && discount.active">
-              <td bitCell colspan="2">
-                <small>
-                  {{ "customBillingStart" | i18n }}
-                  <a routerLink="/settings/subscription/billing-history">
-                    {{ "billingHistory" | i18n }}
-                  </a>
-                  {{ "customBillingEnd" | i18n }}
-                </small>
-              </td>
-            </tr>
           </ng-container>
           <ng-container *ngIf="userOrg.isFreeOrg">
             <tr bitRow *ngIf="userOrg.usePasswordManager">
