+++ apps/web/src/locales/en/messages.json
@@ -7273,12 +7273,6 @@
   "alreadyHaveAccount": {
     "message": "Already have an account?"
   },
-  "customBillingStart": {
-    "message": "Custom billing is not reflected. Visit the "
-  },
-  "customBillingEnd": {
-    "message": " page for latest invoicing."
-  },
   "typePasskey": {
     "message": "Passkey"
   },
