+++ libs/common/src/billing/models/response/organization-subscription.response.ts
@@ -1,9 +1,9 @@
 import { OrganizationResponse } from "../../../admin-console/models/response/organization.response";
+import { BaseResponse } from "../../../models/response/base.response";
 
 import {
   BillingSubscriptionResponse,
   BillingSubscriptionUpcomingInvoiceResponse,
-  BillingCustomerDiscount,
 } from "./subscription.response";
 
 export class OrganizationSubscriptionResponse extends OrganizationResponse {
@@ -11,7 +11,7 @@ export class OrganizationSubscriptionResponse extends OrganizationResponse {
   storageGb: number;
   subscription: BillingSubscriptionResponse;
   upcomingInvoice: BillingSubscriptionUpcomingInvoiceResponse;
-  discount: BillingCustomerDiscount;
+  customerDiscount: BillingCustomerDiscount;
   expiration: string;
   expirationWithoutGracePeriod: string;
   secretsManagerBeta: boolean;
@@ -27,10 +27,24 @@ export class OrganizationSubscriptionResponse extends OrganizationResponse {
       upcomingInvoice == null
         ? null
         : new BillingSubscriptionUpcomingInvoiceResponse(upcomingInvoice);
-    const discount = this.getResponseProperty("Discount");
-    this.discount = discount == null ? null : new BillingCustomerDiscount(discount);
+    const customerDiscount = this.getResponseProperty("CustomerDiscount");
+    this.customerDiscount =
+      customerDiscount == null ? null : new BillingCustomerDiscount(customerDiscount);
     this.expiration = this.getResponseProperty("Expiration");
     this.expirationWithoutGracePeriod = this.getResponseProperty("ExpirationWithoutGracePeriod");
     this.secretsManagerBeta = this.getResponseProperty("SecretsManagerBeta");
+  }
+}
+
+export class BillingCustomerDiscount extends BaseResponse {
+  id: string;
+  active: boolean;
+  percentOff?: number;
+
+  constructor(response: any) {
+    super(response);
+    this.id = this.getResponseProperty("Id");
+    this.active = this.getResponseProperty("Active");
+    this.percentOff = this.getResponseProperty("PercentOff");
   }
 }
